+++ Assets/_SPC/GamePlay/UI/Scripts/HighScoreEntryUI.cs
@@ -9,7 +9,7 @@ namespace _SPC.GamePlay.UI.Scripts.Scripts
         [SerializeField] private TextMeshProUGUI _nameText;
         [SerializeField] private TextMeshProUGUI _scoreText;
 
-        public void Initialize(int rank, string nickname, int score)
+        public void Initialize(int rank, string nickname, long score)
         {
             _rankText.text = rank.ToString();
             _nameText.text = nickname;
